+++ imio/dms/mail/dmsmail.py
@@ -83,6 +83,7 @@ from z3c.form.browser.radio import RadioFieldWidget
 from z3c.form.interfaces import HIDDEN_MODE
 from zope import schema
 from zope.component import adapts
+from zope.component import getAdapter
 from zope.component import getMultiAdapter
 from zope.component import getUtility
 from zope.interface import alsoProvides
@@ -869,13 +870,13 @@ class ImioDmsOutgoingMail(DmsOutgoingMail):
         :param all_done: if True, check if all approvings are done
         :return: boolean
         """
-        annot = get_approval_annot(self)
-        if not annot.get("users", []):
+        approval = getAdapter(self, name="approvals")
+        if not approval.approvers:
             return False
         elif not all_done:
             return True
         else:  # has approvals and all done
-            return annot.get("approval", 0) == 99
+            return approval.current_nb == -1
 
     def wf_conditions(self):
         """Returns the adapter providing workflow conditions"""