+++ CHANGES.rst
@@ -32,6 +32,8 @@ Changelog
   [chris-adam, sgeulette]
 - Added behavior on Configurable POD Templates to set content category on generated documents.
   [chris-adam]
+- Refactor idm.approval annotation.
+  [chris-adam]
 
 3.0 (2021-09-30)
 ----------------